delphinus_demo/requirements.txt
httpx
python-dotenv
rich
deepdiff
google-genai
streamlit
pytest
